requirements.txt
# Runtime dependencies
psycopg2-binary>=2.9.5,<3.0
pandas
openpyxl

# Build dependencies (not needed at runtime)
pyinstaller>=4.7,<5.0